requirements.txt
# Tijan AI — Backend v6.1
# Render Standard (1 CPU / 2GB RAM)
# All versions pinned — do NOT change without running: pytest tests/

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4

# PDF
reportlab==4.1.0
pypdf==4.1.0

# Parsing géométrie
ezdxf==1.4.2

# PDF vectoriel (extraction texte + cotes)
pymupdf==1.25.3

# IA (parsing plans scannés si besoin futur)
anthropic==0.49.0

# Excel + Word output
openpyxl==3.1.2
python-docx==1.1.0

# Utilitaires
python-multipart==0.0.9
requests==2.31.0
httpx==0.27.0
openai==1.66.5

# Testing (dev only — not needed on Render)
# pytest>=8.0
